Add WizardControllerConcern
This adds a concern which handles common logic related to the three
wizard controllers to reduce code duplication.
